requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.25
pymysql==1.1.0
aiomysql==0.2.0
aiokafka==0.12.0
redis==5.0.1
python-dotenv==1.0.0
alembic==1.13.1
pydantic[email]==2.6.1
email-validator==2.1.0.post1
pydantic-settings==2.1.0
